BUG: Rails 3.0 does not respond to config.assets
I spent a few hours with this fix since it's so
hard to work with a faked Rails.application.config.

I finally settled for the easiest fix possible,
even though it does not look good.

This is a good sign I need to start testing the
real integrations using real Rails apps;
preferably with Engines or something similar.