Remove documentation about passing CSS filenames as symbols
This should fix #50 somewhat. In an optimal world, I'd like to actually
behave like Rails 4 does when a symbol is passed, but that would be
backwards-incompatible and I am not ready to release a new major version
yet. Just not talking about it should make it easier for people to
discover the problem themselves and fix it.

In the next major version, this could be a proper feature instead.