Remove direct invocations of Inliner from outside of Document
This code isn't pretty and I did not add a test for it either; the way
it is structured right now is actually very ugly. The current plan is to
start reworking Inliner internals, starting with that horrible
constructor. To do that, I need to reduce the surface which that class
is occupying.

Replacing the helper method is even harder to do since it is stubbed
everywhere and changing the stubs to accomodate the current version of
Document will be hard. I'll get to that problem in due time.