Extract StyleProperties domain object for a collection of StyleProperty
Also adjusted how style properties are being produced. It used to be
that "duplicate" properties were being filtered out (e.g. two "color"
properties resulted on only one "color" property to be outputed), but
I've realized that that was a mistake. Modern CSS is build around
cascading properties as well:

  body { background: #deadbeef; background: -webkit-gradient(lol); background: gradient(lol); }

I cannot assume that some of them should go away. This leads to much
larger HTML documents for those who aren't careful, but it's the only
way to be future-proofed. Users of this gem must take care not to bloat
the stylesheets too much.